src/requirements.txt
amqp==5.2.0
argon2-cffi==21.3.0
asgiref==3.7.2
async-generator==1.10
attrs==23.2.0
backcall==0.2.0
billiard>=4.2.0
bleach==6.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
click==8.1.7
click-didyoumean>=0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
entrypoints==0.4
idna==3.7
inflect==7.0.0
ipykernel==6.29.2
ipython==8.22.1
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonschema==4.21.1
jupyter-client==8.6.0
jupyter-core==5.7.2
jupyterlab==4.0.12
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.2
kiwisolver==1.4.5
kombu==5.3.5
MarkupSafe==2.1.5
matplotlib>=3.8.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.0.8
numpy==1.26.4
packaging==23.2
pandas>=2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.2.0
prometheus-client==0.20.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.2
pyparsing==3.1.1
pyrsistent==0.20.0
python-dateutil==2.9.0
pytz==2024.1
pyzmq==26.0.2
redis==5.0.3
requests==2.31.0
Send2Trash==1.8.3
six==1.16.0
sqlparse==0.2.4
terminado==0.18.1
testpath==0.6.0
tornado==6.4
traitlets==5.14.2
urllib3==2.2.1
uWSGI==2.0.23
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
